Update PlayerInput state
Move state into the PlayerInput's parent in order to better manage state for all players.
